Page modifier une sortie

+++ src/Controller/SortieController.php
@@ -18,6 +18,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class SortieController extends AbstractController
 {
@@ -68,7 +69,7 @@ class SortieController extends AbstractController
 
         if ($idLieu && $submit) {
             $lieu = $lieuRepository->find($idLieu);
-            $etat = ($submit == "enregistrer") ? $etatRepository->find(1) : $etatRepository->find(2);
+            $etat = ($submit == "enregistrer") ? $etatRepository->rechercheParLibelle("Créée") : $etatRepository->rechercheParLibelle("Ouverte");
 
             $sortie->setLieu($lieu);
             $sortie->setEtat($etat);
@@ -128,73 +129,81 @@ class SortieController extends AbstractController
     }
 
     #[Route('/ville/{id}/lieux')]
-public function recupererLieuxDuneVille(LieuRepository $lieuRepository, VilleRepository $villeRepository, int $id)
-{
-    $ville = $villeRepository->find($id);
-    $lieux = $lieuRepository->rechercheParVille($ville);
+    public function recupererLieuxDuneVille(LieuRepository $lieuRepository, VilleRepository $villeRepository, int $id)
+    {
+        $ville = $villeRepository->find($id);
+        $lieux = $lieuRepository->rechercheParVille($ville);
 
-    $lieuxTableau = [];
-    foreach ($lieux as $lieu) {
-        $lieuxTableau[] = [
-            'id' => $lieu->getId(),
-            'nom' => $lieu->getNom(),
-            'rue' => $lieu->getRue(),
-            'codePostal' => $ville->getCodePostal(),
-            'latitude' => $lieu->getLatitude(),
-            'longitude' => $lieu->getLongitude(),
-        ];
+        $lieuxTableau = [];
+        foreach ($lieux as $lieu) {
+            $lieuxTableau[] = [
+                'id' => $lieu->getId(),
+                'nom' => $lieu->getNom(),
+                'rue' => $lieu->getRue(),
+                'codePostal' => $ville->getCodePostal(),
+                'latitude' => $lieu->getLatitude(),
+                'longitude' => $lieu->getLongitude(),
+            ];
+        }
+        return new JsonResponse($lieuxTableau);
     }
-    return new JsonResponse($lieuxTableau);
-}
 
     #[Route('/sortie/{id}/supprimer', name: 'sortie_supprimer')]
-    public function supprimerSortie(int $id, EntityManagerInterface $entityManager): Response
+    public function supprimerSortie(int $id, EntityManagerInterface $entityManager, SortieRepository $sortieRepository): Response
     {
         // Récupérer la sortie depuis la base de données
-        $sortie = $entityManager->getRepository(Sortie::class)->find($id);
+        $sortie = $sortieRepository->find($id);
 
-        // Supprimer la sortie
+        if ($this->getUser() !== $sortie->getOrganisateur()) {
+            throw new AccessDeniedException("Accès interdit. Vous n'êtes pas l'organisateur de cette sortie.");
+        }
+
         $entityManager->remove($sortie);
         $entityManager->flush();
-
-        // Rechargement de la page après la suppression
-        return $this->render('pages/modifierSortie.html.twig');
+        return $this->redirectToRoute('app_accueil');
     }
 
-    #[Route('/sortie/modifier/{id}', name: 'sortie_modifier')]
-    public function modifier(Request $request, EntityManagerInterface $entityManager, LieuRepository $lieuRepository, int $id): Response
+    #[Route('/sortie/{id}/modifier', name: 'sortie_modifier')]
+    public function modifier(Request $request, EntityManagerInterface $entityManager, LieuRepository $lieuRepository, EtatRepository $etatRepository, int $id): Response
     {
+
         $sortie = $entityManager->getRepository(Sortie::class)->find($id);
+
+        if ($this->getUser() !== $sortie->getOrganisateur()) {
+            throw new AccessDeniedException("Accès interdit. Vous n'êtes pas l'organisateur de cette sortie.");
+        }
 
         $sortieForm = $this->createForm(SortieType::class, $sortie);
         $sortieForm->handleRequest($request);
 
         // Récupération de la liste des lieux
         $lieux = $lieuRepository->findAll();
-        $lieuxTableau = [];
-        foreach ($lieux as $lieu) {
-            $ville = $lieu->getVille();
-            $codePostal = $ville ? $ville->getCodePostal() : null;
-            $lieuxTableau[] = [
-                'id' => $lieu->getId(),
-                'nom' => $lieu->getNom(),
-                'rue' => $lieu->getRue(),
-                'codePostal' => $codePostal,
-                'latitude' => $lieu->getLatitude(),
-                'longitude' => $lieu->getLongitude(),
-            ];
-        }
-        $lieuxJson = json_encode($lieuxTableau);
+
 
         if ($sortieForm->isSubmitted() && $sortieForm->isValid()) {
+
+            $idLieu = $request->request->get('lieu', '');
+            $submit = $request->request->get('submit', '');
+
+            if ($idLieu && $submit) {
+                $lieu = $lieuRepository->find($idLieu);
+                dump($lieu);
+                $etat = ($submit == "enregistrer") ? $etatRepository->rechercheParLibelle("Créée") : $etatRepository->rechercheParLibelle("Ouverte");
+
+                $sortie->setLieu($lieu);
+                $sortie->setEtat($etat);
+            }
+
+
+            $entityManager->persist($sortie);
             $entityManager->flush();
             return $this->redirectToRoute('app_accueil');
         }
 
         return $this->render('pages/modifierSortie.html.twig', [
             'sortieForm' => $sortieForm->createView(),
-            'lieux' => $lieuxJson // Passer les données des lieux au template
+            'lieux' => $lieux,
+            'sortie' => $sortie,
         ]);
     }
-
 }
